refactor(model): Move snake and ladder classes to new package structure
