+++ src/main/java/ntnu/idatt2003/file/BoardFileReader.java
@@ -1,7 +1,7 @@
 package ntnu.idatt2003.file;
 
 import java.nio.file.Path;
-import ntnu.idatt2003.model.SnakeLadderBoard;
+import ntnu.idatt2003.model.snakeandladder.SnakeLadderBoard;
 
 public interface BoardFileReader {
   SnakeLadderBoard readBoard(Path path) throws Exception;
