+++ src/main/resources/players.csv
@@ -1,3 +1,3 @@
 Name,Age,Icon
-ra,20CAR
-ma,23CAT
+l,21DOG
+m,23CAT
